implement `cargo sqlx prepare`
also organize code more clearly in `cargo-sqlx`
