+++ cargo-sqlx/src/migrator/sqlite.rs
@@ -9,7 +9,7 @@ use sqlx::SqliteConnection;
 use anyhow::{anyhow, Context, Result};
 use async_trait::async_trait;
 
-use crate::database_migrator::{DatabaseMigrator, MigrationTransaction};
+use crate::migrator::{DatabaseMigrator, MigrationTransaction};
 
 pub struct Sqlite {
     db_url: String,
